Extract fetch from fetchJson
